requirements.txt
requests
beautifulsoup4
phonenumbers
pytz
reverse-geocoder
python-whois
dnspython